+++ custom_components/linknlink/const.py
@@ -4,5 +4,6 @@ from homeassistant.const import Platform
 
 DOMAIN = "linknlink"
 CONF_HOST = "host"
-PLATFORMS = (Platform.SENSOR, Platform.BINARY_SENSOR, Platform.EVENT)
+CONF_LOCAL_KEY = "local_key"
+PLATFORMS = (Platform.SENSOR, Platform.BINARY_SENSOR, Platform.EVENT, Platform.SWITCH)
 UPDATE_INTERVAL_SECONDS = 30